Reader writer refactor and improve performance (~3.6x faster)

+++ utils/getDirSize.go
@@ -6,13 +6,13 @@ import (
 )
 
 // walk a directory and get the total file size
-func GetDirSize(path string) (size int64, err error) {
+func GetDirSize(path string) (size int, err error) {
 	err = filepath.Walk(path, func(path string, info fs.FileInfo, err error) error {
 		if err != nil {
 			return err
 		}
 		if !info.IsDir() {
-			size += info.Size()
+			size += int(info.Size())
 		}
 		return nil
 	})
